pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-sre-env"
version = "0.1.0"
description = "SRE Incident Investigation environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "uvicorn>=0.29.0",
    "fastapi>=0.110.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
inference = [
    "openai>=1.0.0",
]

[project.scripts]
server = "sre_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["sre_env", "sre_env.server", "sre_env.tasks"]
package-dir = { "sre_env" = ".", "sre_env.server" = "server", "sre_env.tasks" = "tasks" }
